Refactor job update and delete routes
Cleaned up the job update and delete routes in the web.php file. Removed unnecessary code and added validation for the title and description fields in the update route. Also added authorization checks, which are currently on hold. In the delete route, added authorization checks and deleted the job with the specified ID. Redirects to the appropriate pages after successful update or delete.